+++ db/migrate/20250224212916_add_progress_fields_to_training_memberships.rb
@@ -1,0 +1,9 @@
+class AddProgressFieldsToTrainingMemberships < ActiveRecord::Migration[7.2]
+  def change
+    add_column :training_memberships, :completed_at, :datetime
+    add_column :training_memberships, :completion_percentage, :integer, default: 0, null: false
+
+    add_index :training_memberships, :completed_at
+    add_index :training_memberships, :completion_percentage
+  end
+end